REFACTOR: 파일Rename, 불필요한 파일 제거 및  경로 수정
