+++ src/components/my-page/my-activity/MyPostFrame.tsx
@@ -4,7 +4,7 @@ import LinearGradient from 'react-native-linear-gradient';
 
 import LocationGreyIcon from 'assets/image/icon/list_location.svg';
 import FavoriteButton from 'components/reuse/button/FavoriteButton';
-import { LikeReview } from 'hooks/axios/ReviewDetail';
+import { LikeReview } from 'hooks/axios/Review';
 import { storage } from 'hooks/mmkv/storage';
 import { useAppSelector } from 'hooks/redux/store';
 import { ReviewFrameProps } from 'interfaces/Home.interface';
