+++ src/hooks/axios/Event.ts
@@ -1,6 +1,21 @@
 import { axiosInstance } from './ApiHeader';
 
-//TODO: 테스트 필요함
+export const GetEventData = async (event_Id: number) => {
+    return await axiosInstance({
+        method: 'get',
+        url: `/events/${event_Id}`,
+        params: {
+            event_Id,
+        },
+    })
+        .then(res => {
+            return res.data;
+        })
+        .catch(error => {
+            console.log(error);
+        });
+};
+
 export const LikeEvent = async (event_Id: number | null | undefined) => {
     return await axiosInstance({
         method: 'post',
