--- src/hooks/axios/EventDetail.ts
@@ -1,17 +1,0 @@
-import { axiosInstance } from './ApiHeader';
-
-export const GetEventData = async (event_Id: number) => {
-    return await axiosInstance({
-        method: 'get',
-        url: `/events/${event_Id}`,
-        params: {
-            event_Id,
-        },
-    })
-        .then(res => {
-            return res.data;
-        })
-        .catch(error => {
-            console.log(error);
-        });
-};
