+++ src/hooks/axios/SearchLocation.ts
@@ -1,13 +1,5 @@
 import { axiosInstance } from './ApiHeader';
 
-//TODO: latitude, longitude, radius 가 backend에서 수정될 예정, 추후 변경
-
-/**
- * Test
- * photo_booth_brand_name: string
- * latitude: 36.8101475281
- * longitude: 127.1470316068
- * */
 export const GetLocationSearchData = async (
     photo_booth_brand_name: string | null,
     latitude: number | null,
@@ -20,8 +12,8 @@ export const GetLocationSearchData = async (
 
         params: {
             photo_booth_brand_name,
-            latitude: 36.8101475281,
-            longitude: 127.1470316068,
+            latitude,
+            longitude,
             radius,
         },
     })
